requirements.txt
openai-whisper
requests
PyYAML
schedule
sounddevice
numpy
flask
